+++ src/models/container/MHBucketPagedResponse.js
@@ -1,0 +1,13 @@
+import MHPagedResponse from './MHPagedResponse.js';
+
+export default class MHBucketPagedResponse extends MHPagedResponse {
+  get jsonProperties() {
+    return {
+      ...super.jsonProperties,
+      type: String,
+      name: String,
+      formattedName: String,
+      hintText: String,
+    };
+  }
+}