+++ src/models/container/MHContext.js
@@ -1,5 +1,6 @@
 import { jsonCreateWithArgs } from '../internal/jsonParse.js';
 import MHRelationship from './MHRelationship.js';
+import MHRelationalPair from './MHRelationalPair.js';
 import MHSorting from './MHSorting.js';
 import MHSourceMedium from '../source/MHSourceMedium.js';
 
@@ -12,8 +13,10 @@ export default class MHContext {
     return {
       consumable: Boolean,
       sorting: MHSorting,
+      fixed: MHSorting,
       relationships: [MHRelationship],
-      mediums: [MHSourceMedium]
+      mediums: [MHSourceMedium],
+      difference: [MHRelationalPair]
     };
   }
 }